+++ modules/tinymce/src/core/test/ts/atomic/table/TableUtilsTest.ts
@@ -2,9 +2,9 @@ import { describe, it } from '@ephox/bedrock-client';
 import { assert } from 'chai';
 import * as fc from 'fast-check';
 
-import { addPxSuffix, isPercentage, isPixel, removePxSuffix } from 'tinymce/plugins/table/core/Util';
+import { isPercentage, isPixel } from 'tinymce/core/table/TableUtils';
 
-describe('atomic.tinymce.plugins.table.core.UtilTest', () => {
+describe('atomic.tinymce.core.table.TableUtilsTest', () => {
   it('isPercentage', () => {
     assert.isFalse(isPercentage(''), 'Empty string is false');
     assert.isFalse(isPercentage('%'), 'Single % string is false');
@@ -30,29 +30,6 @@ describe('atomic.tinymce.plugins.table.core.UtilTest', () => {
       assert.isFalse(isPixel(n + ''), 'Number string is false');
       assert.isFalse(isPixel(n + '%'), 'Percent string is false');
       assert.isFalse(isPixel(n + 'px' + n), 'String containing px string is false');
-    }));
-  });
-
-  it('removePxSuffix', () => {
-    assert.equal(removePxSuffix(''), '', 'Empty string is identical');
-    assert.equal(removePxSuffix('10px'), '10', 'Pixel string has pixel removed');
-
-    fc.assert(fc.property(fc.float(1, 100), (n) => {
-      assert.equal(removePxSuffix(n + 'px'), n + '', 'Arbitrary float with px string is true');
-      assert.equal(removePxSuffix(n + ''), n + '', 'Number string is identical');
-      assert.equal(removePxSuffix('px' + n), 'px' + n, 'String with pixel prefix is identical');
-      assert.equal(removePxSuffix(n + '%'), n + '%', 'Percent string is identical');
-    }));
-  });
-
-  it('addPxSuffix', () => {
-    assert.equal(addPxSuffix(''), '', 'Empty string is identical');
-    assert.equal(addPxSuffix('10'), '10px', 'Number string has px added');
-
-    fc.assert(fc.property(fc.float(1, 100), (n) => {
-      assert.equal(addPxSuffix(n + ''), n + 'px', 'Arbitrary float with px string is true');
-      assert.equal(addPxSuffix(n + '%'), n + '%', 'Percent string is identical');
-      assert.equal(addPxSuffix(n + 'px'), n + 'px', 'Pixel string is identical');
     }));
   });
 });