+++ modules/tinymce/src/core/test/ts/browser/init/InitEditorNoThemeInlineTest.ts
@@ -1,4 +1,4 @@
-import { Assertions } from '@ephox/agar';
+import { ApproxStructure, Assertions } from '@ephox/agar';
 import { describe, it } from '@ephox/bedrock-client';
 import { SelectorFind, SugarBody, Traverse } from '@ephox/sugar';
 import { TinyAssertions, TinyDom, TinyHooks } from '@ephox/wrap-mcagar';
@@ -26,13 +26,17 @@ describe('browser.tinymce.core.init.InitEditorNoThemeInlineTest', () => {
     const editor = hook.editor();
     const body = SugarBody.body();
     const targetElement = SelectorFind.descendant(body, '#' + editor.id).getOrDie('No elm');
-    const nextElement = Traverse.nextSibling(targetElement);
+    const nextElement = Traverse.nextSibling(targetElement).getOrDie('Should be an element after the target');
 
     // TODO FIXME this seems like an odd exception
     assert.isNull(editor.getContainer(), 'Should be null since inline without a theme does not set editorContainer');
     Assertions.assertDomEq('Should be expected editor body element', targetElement, TinyDom.body(editor));
     Assertions.assertDomEq('Should be expected editor target element', targetElement, TinyDom.targetElement(editor));
     Assertions.assertDomEq('Editor.contentAreaContainer should equal target element', targetElement, TinyDom.contentAreaContainer(editor));
-    assert.isTrue(nextElement.isNone(), 'Should be no element after target');
+    // The only element that should be after the target is the TableResizeHandler TableWire div element
+    assert.lengthOf(Traverse.nextSiblings(targetElement), 1, 'Should only be one element after target');
+    // test with 5px
+    const resizeDivStructure = ApproxStructure.fromHtml('<div style="position: static; height: 0px; width: 0px; padding: 0px; margin: 0px; border: 0px;"></div>');
+    Assertions.assertStructure('Sibling element should be a resize div', resizeDivStructure, nextElement);
   });
 });