+++ modules/tinymce/src/plugins/table/test/ts/browser/NewCellRowEventsTest.ts
@@ -5,6 +5,8 @@ import { assert } from 'chai';
 import Editor from 'tinymce/core/api/Editor';
 import Plugin from 'tinymce/plugins/table/Plugin';
 
+import { insertTable } from '../module/test/TableTestUtils';
+
 describe('browser.tinymce.plugins.table.NewCellRowEventsTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
     plugins: 'table',
@@ -15,23 +17,23 @@ describe('browser.tinymce.plugins.table.NewCellRowEventsTest', () => {
     base_url: '/project/tinymce/js/tinymce'
   }, [ Plugin ]);
 
-  it('TBA: Table newcell/newrow events', () => {
+  it('TBA: Table NewCell/NewRow events', () => {
     const editor = hook.editor();
     const cells: HTMLTableCellElement[] = [];
     const rows: HTMLTableRowElement[] = [];
     let counter = 0;
 
-    editor.on('newcell', (e) => {
+    editor.on('NewCell', (e) => {
       cells.push(e.node);
-      e.node.setAttribute('data-counter', counter++);
+      e.node.setAttribute('data-counter', '' + counter++);
     });
 
-    editor.on('newrow', (e) => {
+    editor.on('NewRow', (e) => {
       rows.push(e.node);
-      e.node.setAttribute('data-counter', counter++);
+      e.node.setAttribute('data-counter', '' + counter++);
     });
 
-    editor.plugins.table.insertTable(2, 3);
+    insertTable(editor, { rows: 3, columns: 2 });
 
     assert.lengthOf(cells, 6);
     assert.lengthOf(rows, 3);